+++ webpack.config.js
@@ -1,0 +1,16 @@
+/**
+ * WordPress Dependencies
+ */
+const defaultConfig = require( '@wordpress/scripts/config/webpack.config' );
+const path = require( 'path' );
+
+module.exports = {
+	...defaultConfig,
+	entry: {
+		index: path.resolve( process.cwd(), 'blocks/eventbrite-widget/src', 'index.js' ),
+	},
+	output: {
+		filename: '[name].js',
+		path: path.resolve( process.cwd(), 'blocks/eventbrite-widget/build' ),
+	},
+};